+++ test/integration/test_cases/web_content/cookbook/exports/animated_gif_export.mjs
@@ -17,8 +17,8 @@ const testSteps = [
       workers: 4,
       workerScript: URL.createObjectURL(workerBlob),
       quality: 10,
-      width: chart.feature.htmlCanvas.element().width,
-      height: chart.feature.htmlCanvas.element().height
+      width: chart.feature.htmlCanvas.element.width,
+      height: chart.feature.htmlCanvas.element.height
     })
 
     chart.on('draw-complete', (event) => {