+++ test/integration/test_cases/web_content/cookbook/interactive/mouse_scroll.mjs
@@ -61,8 +61,8 @@ const testSteps = [
     })
 
     chart.on('pointerdown', (event) => {
-      const convert = chart.getConverter('plot-area', 'canvas', 'relative')
-      const rel = convert(event.detail.position)
+      const coordSystem = chart.feature.coordSystem
+      const rel = coordSystem.toRelative(event.detail.position)
       scroller.drag(rel.x)
       event.preventDefault()
     })
@@ -73,8 +73,8 @@ const testSteps = [
     })
 
     chart.on('pointermove', (event) => {
-      const convert = chart.getConverter('plot-area', 'canvas', 'relative')
-      const rel = convert(event.detail.position)
+      const coordSystem = chart.feature.coordSystem
+      const rel = coordSystem.toRelative(event.detail.position)
       scroller.track(rel.x)
       throttle.call(() =>
         chart.animate(