+++ CHANGELOG.md
@@ -37,7 +37,6 @@
   triggered on.
 - Drawing events are containing drawing geometry in an improved format.
 - Missing TypeScript definitions for Event objects added.
-- Converter functions to convert between canvas and relative coordinates.
 - Orientation new default parameter: `auto`.
 - New dataseries `unit` parameter introduced, undocumented unit parsing from 
   parameter name string removed.
@@ -55,8 +54,10 @@
 - TypeScript interface made more typesafe, while the javascript api got ported to 
 - TypeScript.
 - `pointeron` event removed, `pointerleave` event added. 
-- `Vizzu.feature.htmlCanvas.element()` introduced to expose the underlying HTML canvas element.
+- `Vizzu.feature.htmlCanvas.element` introduced to expose the underlying HTML canvas element.
 - `Vizzu.feature.rendering.update()` introduced to trigger chart re-rendering.
+- `Vizzu.feature.coordSystem.toCanvas()`/`toRelative()` converter functions to convert
+  between canvas and relative coordinates.
 
 
 ## [0.8.1] - 2023-08-15